// frontend/src/enums/CountryEnum.ts
export enum CountryEnum {
    ALBANIA = 1,
    ANDORRA = 2,
    AUSTRIA = 3,
    BELARUS = 4,
    BELGIUM = 5,
    BOSNIA_AND_HERZEGOVINA = 6,
    BULGARIA = 7,
    CROATIA = 8,
    CYPRUS = 9,
    CZECH_REPUBLIC = 10,
    DENMARK = 11,
    ESTONIA = 12,
    FINLAND = 13,
    FRANCE = 14,
    GEORGIA = 15,
    GERMANY = 16,
    GREECE = 17,
    HUNGARY = 18,
    ICELAND = 19,
    IRELAND = 20,
    ITALY = 21,
    KAZAKHSTAN = 22,
    KOSOVO = 23,
    LATVIA = 24,
    LIECHTENSTEIN = 25,
    LITHUANIA = 26,
    LUXEMBOURG = 27,
    MALTA = 28,
    MOLDOVA = 29,
    MONACO = 30,
    MONTENEGRO = 31,
    NETHERLANDS = 32,
    NORTH_MACEDONIA = 33,
    NORWAY = 34,
    POLAND = 35,
    PORTUGAL = 36,
    ROMANIA = 37,
    RUSSIA = 38,
    SAN_MARINO = 39,
    SERBIA = 40,
    SLOVAKIA = 41,
    SLOVENIA = 42,
    SPAIN = 43,
    SWEDEN = 44,
    SWITZERLAND = 45,
    TURKEY = 46,
    UKRAINE = 47,
    UNITED_KINGDOM = 48,
    VATICAN_CITY = 49,
}

export namespace CountryEnum {
    export const getOptions = (t: (key: string) => string): { value: number; label: string }[] => {
        return Object.entries(CountryEnum)
            .filter(([_, value]) => typeof value === 'number')
            .map(([key, value]) => ({
                value: value as number,
                label: t(`countries.${key.toLowerCase()}`),
            }));
    };
}